+++ Hampi/HMPFirebaseKit/HMPFirebaseKit.h
@@ -1,0 +1,19 @@
+//
+//  HMPFirebaseKit.h
+//  HMPFirebaseKit
+//
+//  Created by Joan Molinas Ramon on 25/5/17.
+//  Copyright © 2017 com.hamp. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for HMPFirebaseKit.
+FOUNDATION_EXPORT double HMPFirebaseKitVersionNumber;
+
+//! Project version string for HMPFirebaseKit.
+FOUNDATION_EXPORT const unsigned char HMPFirebaseKitVersionString[];
+
+#import <HMPFirebaseKit/HMPFAuth.h>
+
+